support using tarball source for package

+++ TBD.sh
@@ -5,14 +5,26 @@ source $1
 # update git cache
 cachedir=${TBD_CACHE:-$(realpath $(dirname $1))/.cache}
 mkdir -p $cachedir
-if [ -d $cachedir/$pkgname ]; then
-  opwd=$PWD
+opwd=$PWD
+if [ ! -z "$pkggit" ]; then
+  if [ -d $cachedir/$pkgname ]; then
+    cd $cachedir/$pkgname
+    git reset --hard HEAD
+    git fetch --all
+    cd $opwd
+  else
+    git clone $pkggit $cachedir/$pkgname
+  fi
+elif [ ! -z "$pkgurl" ]; then
+  wget $pkgurl -O /tmp/$pkgname.tbd-downloaded
+  
+  mkdir -p $cachedir/$pkgname
   cd $cachedir/$pkgname
-  git reset --hard HEAD
-  git fetch --all
+  tar xvf /tmp/$pkgname.tbd-downloaded
   cd $opwd
 else
-  git clone $pkggit $cachedir/$pkgname
+  echo "Either pkggit or pkgurl must be set in the .TBD file."
+  exit 127
 fi
 
 CRI="${TBD_CRI:-docker}"
@@ -22,8 +34,8 @@ if [ "$(tty)" != "not a tty" ]; then
 fi
 
 if [ $CRI == "none" ]; then
-  local pwd=$(dirname "$(realpath "${BASH_SOURCE[0]}")")
-  export BUILD_DIR=$pwd/$cachedir/$pkgname
+  pwd=$(dirname "$(realpath "${BASH_SOURCE[0]}")")
+  export BUILD_DIR=$cachedir/$pkgname
   source $1
   source $pwd/iTBD.sh
 else
